PLAN
Plan of tests that will need to be implemented. Every DAG is a Binary Tree. Some changes will need to be made, For personal use and to distinguish between my BT solution I will change key's from Integers to Strings. From first glance it seems an Array List would be the best way to display this information.