Adding More Tests For FindLCA
The user cannot input non existing nodes. I need to figure out how to test for this/ if it needs to be tested for. 
Next step is to make sure the graph is acyclic, the basics of findLCA are done and I now need to look for ways in which the user could "break" this code and test for it.